Remove old backend and rename fastapi
