+++ backend/.env.example
@@ -1,0 +1,1 @@
+SQLALCHEMY_DATABASE_URL=sqlite:///onacut.data
