--- backend/onacut/api/district/fields.py
@@ -1,10 +1,0 @@
-from marshmallow import Schema, fields
-
-# from flask_restful import fields
-
-
-class DistrictGetResponseSchema(Schema):
-    id = fields.Integer()
-    name = fields.String()
-    city_id = fields.Function(lambda x: x.city.id)
-    num_alerts = fields.Function(lambda x: len(x.alerts))
