--- backend/onacut/api/locations/fields.py
@@ -1,10 +1,0 @@
-from marshmallow import Schema, fields
-
-
-class LocationGetResponseSchema(Schema):
-    id = fields.Integer()
-    name = fields.String()
-    asciiname = fields.String()
-    alternativenames = fields.String()
-    longitude = fields.String()
-    lattitude = fields.String()
