+++ backend/requirements-dev.txt
@@ -1,7 +1,13 @@
-# for tests
-pytest
-
-# for lint
-isort
-black
-flake8
+fastapi==0.85.0
+SQLAlchemy==1.4.41
+starlette==0.20.4
+requests==2.28.1
+pydantic==1.10.2
+pydantic-choices==0.1.1
+pytest==7.1.3
+pytest-mock==3.10.0
+uvicorn==0.18.3
+black==22.10.0
+isort==5.10.1
+flake8==5.0.4
+python-dotenv==0.21.0
