+++ backend/requirements.txt
@@ -1,10 +1,6 @@
-Faker==13.3.0
-Flask==2.0.3
-Flask-Migrate==3.1.0
-Flask-RESTful==0.3.9
-flask-apispec==0.11.0
-Flask-SQLAlchemy==2.5.1
-Flask-Cors==3.0.10
-werkzeug>=2
-psycopg2-binary==2.9.3
-python-dotenv==0.19.2
+fastapi==0.85.0
+SQLAlchemy==1.4.41
+pydantic==1.10.2
+pydantic-choices==0.1.1
+uvicorn==0.18.3
+python-dotenv==0.21.0
