--- backend/server.py
@@ -1,4 +1,0 @@
-from onacut import app
-
-if __name__ == "__main__":
-    app.run("0.0.0.0")
