--- backend/tests/test_simple.py
@@ -1,6 +1,0 @@
-""" simple assert module """
-
-
-def test_simple():
-    """simple assert"""
-    assert "VsCode" != "Vim"
