--- fastapi/.env.example
@@ -1,1 +1,0 @@
-SQLALCHEMY_DATABASE_URL=sqlite:///onacut.data
