--- fastapi/requirements.txt
@@ -1,6 +1,0 @@
-fastapi==0.85.0
-SQLAlchemy==1.4.41
-pydantic==1.10.2
-pydantic-choices==0.1.1
-uvicorn==0.18.3
-python-dotenv==0.21.0
